+++ src/components/login/EmailLogin.jsx
@@ -90,25 +90,7 @@ const EmailLogin = () => {
   const onMoalClose = () => {
     setOpen({ isOpen: false });
   };
-  let eventSource;
-  const fetchSse = async () => {
-    try {
-      //EventSource생성.
-      eventSource = new EventSourcePolyfill("https://eb.jxxhxxx.shop/connect", {
-        //headers에 토큰을 꼭 담아줘야 500이 안뜬다.
-        headers: {
-          Authorization: sessionStorage.getItem("accessToken"),
-        },
-        withCredentials: true,
-      });
-      // SSE 연결 성공 시 호출되는 이벤트 핸들러
-      eventSource.onopen = () => {
-        console.log("SSE 연결완료");
-      };
-    } catch (error) {
-      console.log("에러발생:", error);
-    }
-  };
+
   //서버에 전달
   const onSubmit = async (event) => {
     event.preventDefault();
@@ -138,7 +120,7 @@ const EmailLogin = () => {
       return;
     }
 
-    await dispatch(
+    dispatch(
       __emailLogin({
         email: emailValue,
         password: passwordValue,
@@ -147,7 +129,7 @@ const EmailLogin = () => {
         setModalStr,
       })
     );
-    onSubmit().then(fetchSse());
+    onSubmit();
   };
 
   //카카오 로그인
